Add CharacterBuilder component
The CharacterBuilder component has a list of character building blocks
and a list of matching characters. Move character lookup logic from App
into CharacterBuilder.
